# vityaz-main/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "operationId": "AuthController_register",
                "summary": "Регистрация пользователя в системе",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "description": "Реквизиты пользователя для регистрации",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterDto"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "BadRequestException": {
                                        "description": "Пользователь с таким адресом электронной почты уже существует",
                                        "value": {
                                            "statusCode": 400,
                                            "message": "User with this email exists",
                                            "error": "Bad Request"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Пользователь с таким адресом электронной почты уже существует",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User with this email exists"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "tags": [
                    "Аутентификация"
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "operationId": "AuthController_login",
                "summary": "Авторизация пользователя в системе или отправка кода",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "description": "Реквизиты пользователя для авторизации",
                    "content": {
                        "application/json": {
                            "schema": {
                                "anyOf": [
                                    {
                                        "$ref": "#/components/schemas/EmailLoginDto"
                                    },
                                    {
                                        "$ref": "#/components/schemas/TraditionalLoginDto"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пользователь успешно авторизовался в системе и получил пару токенов (Работает в случае использования традиционной авторизации)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthTokensDto"
                                }
                            }
                        }
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "UnauthorizedException #1": {
                                        "description": "Неправильно введен адрес электронной почты",
                                        "value": {
                                            "statusCode": 401,
                                            "message": "Incorrect email",
                                            "error": "Unauthorized"
                                        }
                                    },
                                    "UnauthorizedException #2": {
                                        "description": "Неправильный пароль пользователя",
                                        "value": {
                                            "statusCode": 401,
                                            "message": "Incorrect password",
                                            "error": "Unauthorized"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Неправильно введен адрес электронной почты",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Incorrect email"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "tags": [
                    "Аутентификация"
                ]
            }
        },
        "/api/auth/login/confirm": {
            "post": {
                "operationId": "AuthController_confirm",
                "summary": "Подтверждение входа через отправленный код",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfirmationDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пользователь успешно авторизовался в системе и получил пару токенов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthTokensDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "BadRequestException": {
                                        "description": "Пользователь ввел неверный код подтверждения",
                                        "value": {
                                            "statusCode": 400,
                                            "message": "Invalid confirmation code",
                                            "error": "Bad Request"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Пользователь ввел неверный код подтверждения",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid confirmation code"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    },
                    "405": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "MethodNotAllowedException": {
                                        "description": "Пользователю не было отправлено писем",
                                        "value": {
                                            "statusCode": 405,
                                            "message": "User does not have confirmation mails",
                                            "error": "Method Not Allowed"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Пользователю не было отправлено писем",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 405
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User does not have confirmation mails"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Method Not Allowed"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "tags": [
                    "Аутентификация"
                ]
            }
        },
        "/api/auth/refresh": {
            "get": {
                "operationId": "AuthController_refreshTokens",
                "summary": "Продление сессии пользователя",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer refresh token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь успешно продлил свою сессию и получил новую пару токенов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthTokensDto"
                                }
                            }
                        }
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "UnauthorizedException": {
                                        "description": "Пользователь поставил некорректный токен обновления сессии",
                                        "value": {
                                            "statusCode": 401,
                                            "message": "Invalid refresh token",
                                            "error": "Unauthorized"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Пользователь поставил некорректный токен обновления сессии",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid refresh token"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "tags": [
                    "Аутентификация"
                ]
            }
        }
    },
    "info": {
        "title": "Магазин «Витязь»",
        "description": "",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "RegisterDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "smoothbronx@xenofium.com",
                        "writeOnly": true
                    },
                    "password": {
                        "type": "string",
                        "example": "SomePassword1",
                        "writeOnly": true
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
            "EmailLoginDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "description": "Адрес электронной почты пользователя"
                    }
                },
                "required": [
                    "email"
                ]
            },
            "TraditionalLoginDto": {
                "type": "object",
                "properties": {
                    "string": {
                        "type": "string",
                        "example": "smoothbronx@xenofium.com",
                        "writeOnly": true
                    },
                    "password": {
                        "type": "string",
                        "example": "SomePassword1",
                        "writeOnly": true
                    }
                },
                "required": [
                    "string",
                    "password"
                ]
            },
            "AuthTokensDto": {
                "type": "object",
                "properties": {
                    "accessToken": {
                        "type": "string",
                        "readOnly": true
                    },
                    "refreshToken": {
                        "type": "string",
                        "readOnly": true
                    }
                },
                "required": [
                    "accessToken",
                    "refreshToken"
                ]
            },
            "ConfirmationDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "description": "Электронный адрес пользователя",
                        "writeOnly": true
                    },
                    "code": {
                        "type": "number",
                        "description": "Код, прилетевший на его почту",
                        "writeOnly": true
                    }
                },
                "required": [
                    "email",
                    "code"
                ]
            }
        }
    }
}